fix arrays in multi returns

+++ tests/pyccel/test_pyccel.py
@@ -658,7 +658,7 @@ def test_return_numpy_arrays(language):
     types += [float]*2 # 2 floats for c
     types += [complex]*2 # 2 complexs for e
     types += [bool]*2 # 2 bools for d
-    types += [int]*5 # 5 floats for h
+    types += [float]*5 # 5 floats for h
     types += [int]*5 # 5 ints for g
     pyccel_test("scripts/return_numpy_arrays.py", language=language, output_dtype=types)
